datasource_service/pyproject.toml
[tool.poetry]
name = "datasource_service"
version = "0.1.0"
description = "data source api for naxradaws s3 objects"
authors = ["vinayakgadag"]

[tool.poetry.dependencies]
python = "3.9.10"
fastapi = "^0.73.0"
uvicorn = "^0.17.1"
nexradaws = "^1.1"
numpy = "^1.22.1"
scipy = "^1.7.3"
matplotlib = "^3.5.1"
netCDF4 = "^1.5.8"
arm-pyart = "^1.11.6"
pandas = "^1.4.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core==1.1.12"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "datasource_service.main:start"
